+++ app/controllers/v1/slots_controller.rb
@@ -21,14 +21,9 @@ module V1
     # GET /v1/slots/demo
     def show_last
       authorize :stdSlot
-      begin
-        @slots = StdSlotPublic.last(85)
-        @slots += ReSlot.last(15)
-      rescue => ex
-        Airbrake.notify(ex, error_message: "demo slots couldn't be collected")
-      ensure
-        @slots = StdSlotPublic.last(25)
-      end
+      slot_count = ENV['DEMO_SLOTS_COUNT'].nil? ? 100 : ENV['DEMO_SLOTS_COUNT'].to_i
+      @slots = StdSlotPublic.last(slot_count)
+
       render :index
     end
 
@@ -82,12 +77,23 @@ module V1
     # POST /v1/reslot
     def create_webslot
       authorize :stdSlot
+
+      # TODO: we need an unique identifier from each crawler slot to re-identify
+      # exist reslots
+      # Check if Slot already exist
+      user_reslots = current_user.re_slots.unscoped.joins(:meta_slot)
+      same_reslot = user_reslots.where(
+        'meta_slots.start_date = ? AND meta_slots.title = ?',
+        params.require(:startDate), params.require(:title)
+      )
+      return head 421 if same_reslot.any?
+
       # Set Slot Creator:
       slot_creator = User.find_by(email: 'info@timeslot.com')
       # Create MetaSlot:
       metaslot = MetaSlot.find_by(creator_id: slot_creator.id,
-                                  start_date: params.require(:startDate),
-                                  title: params.require(:title))
+                                  start_date: params[:startDate],
+                                  title: params[:title])
       # Create BaseSlot:
       if metaslot
         @slot = BaseSlot.find_by(meta_slot_id: metaslot.id)
@@ -271,6 +277,14 @@ module V1
       authorize @slot
 
       render :comments
+    end
+
+    # GET /v1/slots/1/slotters
+    def show_slotters
+      @slot = BaseSlot.get(params[:id])
+      authorize @slot
+
+      render :slotters
     end
 
     # GET /v1/slots/1/history